requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.2
amqp==5.2.0
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
billiard==4.2.0
black==23.1.0
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.0
celery==5.3.6
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.2.0
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Minify==0.41
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
google-ai-generativelanguage==0.4.0
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-generativeai==0.3.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
htmlmin==0.1.12
httplib2==0.21.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
kombu==5.3.4
lesscpy==0.15.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.26.5
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.3.0
ply==3.11
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.12
psycopg==3.1.15
psycopg-binary==3.1.15
psycopg-pool==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.1
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.1.0
wcwidth==0.2.12
Werkzeug==2.3.6
Whoosh==2.7.4
WTForms==3.0.1
xxhash==3.2.0
yarl==1.8.2
